[57929] feat: refactor footer-sum to JS file, cleanup
