+++ src/main/webapp/WEB-INF/classes/text-webjet9.properties
@@ -2205,4 +2205,4 @@ components.forum.prepare_structure_err.title=Chyba pri príprave štruktúry
 components.forum.prepare_structure_err.text=Pri príprave štruktúry fóra sa vyskytla chyba.
 components.forum.structure=Štruktúra
 
-components.summary.total_title=Celkový počet
+components.summary.total_title=Spolu
