+++ src/main/webapp/WEB-INF/classes/text_cz-webjet9.properties
@@ -2205,4 +2205,4 @@ components.forum.prepare_structure_err.title=Chyba při přípravě struktury
 components.forum.prepare_structure_err.text=Při přípravě struktury fóra se vyskytla chyba.
 components.forum.structure=Struktura
 
-components.summary.total_title=Celkový počet
+components.summary.total_title=Spolu
