+++ src/main/webapp/WEB-INF/classes/text_en-webjet9.properties
@@ -2205,4 +2205,4 @@ components.forum.prepare_structure_err.title=Error preparing structure
 components.forum.prepare_structure_err.text=An error occurred while preparing the forum structure.
 components.forum.structure=Structure
 
-components.summary.total_title=Total count
+components.summary.total_title=Total
