+++ src/main/webapp/apps/stat/admin/index.html
@@ -44,7 +44,8 @@
                 },
                 summary: {
                     mode: "all",
-                    columns: ["visits", "sessions", "uniqueUsers"]
+                    columns: ["visits", "sessions", "uniqueUsers"],
+                    title: "[[#{components.summary.total_title}]]",
                 }
             });
