+++ src/test/webapp/pages/DT.js
@@ -298,15 +298,13 @@ module.exports = {
     },
 
     checkFooterSumValues(I, tableId, values) {
-        within("#" + tableId + "_wrapper", () => {
-            for(let i = 0; i < values.length; i++) {
-                let postfix = "";
-                if(values[i] != null && values[i].length > 0) {
-                    //element "b" is present only if there is value to show
-                    postfix = " > b";
-                }
-                I.see(values[i] + "", ".dt-scroll-footInner > table > tfoot > tr > td:nth-child(" + (i + 1) + ")" + postfix );
+        for(let i = 0; i < values.length; i++) {
+            let postfix = "";
+            if(values[i] != null && values[i].length > 0) {
+                //element "b" is present only if there is value to show
+                postfix = " > b";
             }
-        });
+            I.see(values[i] + "", "#" + tableId + "_wrapper .dt-scroll-footInner > table > tfoot > tr > td:nth-child(" + (i + 1) + ")" + postfix );
+        }
     }
 }
